requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mistune==0.8.4
more-itertools==8.8.0
pymongo==3.12.0
pyrsistent==0.18.0
python-dotenv==0.18.0
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
